requirements.txt
absl-py==2.2.2
alembic==1.14.0
altgraph==0.17.4
astor==0.8.1
attrs==25.1.0
Automat==25.4.16
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==1.5.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
construct==2.10.70
contourpy==1.3.2
cryptography==44.0.1
cycler==0.12.1
docutils==0.21.2
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
gast==0.6.0
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
html5lib==0.9999999
hyperlink==21.0.0
id==1.5.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
joblib==1.5.0
keyring==25.6.0
kiwisolver==1.4.8
macholib==1.16.3
Mako==1.3.6
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.10.3
mdurl==0.1.2
modulegraph==0.19.6
more-itertools==10.7.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
nh3==0.2.21
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.2.1
protobuf==6.31.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py2app==0.28.8
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pyloco==0.0.139
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
pysnmp==7.1.20
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
setuptools==75.8.0
SimpleWebSocketServer==0.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
tensor==0.3.6
tensorboard==1.8.0
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.8.0-py3-none-any.whl#sha256=373c637f9c4f06346b1b02e280b30517d262a1651b252ff889f7c22716ef4548
termcolor==3.1.0
threadpoolctl==3.6.0
trio==0.29.0
trio-websocket==0.12.2
twine==6.1.0
Twisted==24.11.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
ushlex==0.99.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
zope.interface==7.2
gunicorn
python-dotenv